+++ MatrixPilotFBW/airspeedCntrlFBW.c
@@ -33,13 +33,13 @@ int desiredSpeed = (DESIRED_SPEED*10) ;
 
 
 // Calculate the target airspeed in cm/s from desiredSpd in dm/s
-extern int calc_target_airspeed(int desiredSpd, unsigned int airspeed, unsigned int groundspeed);
+extern int16_t calc_target_airspeed(int16_t desiredSpd, uint16_t airspeed, uint16_t groundspeed);
 
 // Calculate the airspeed error vs target airspeed including filtering
-extern int calc_airspeed_error(void);
+extern int16_t calc_airspeed_error(void);
 
 // Calculate the airspeed error integral term with filtering and limits
-extern long calc_airspeed_int_error(int aspdError, long aspd_integral);
+extern int32_t calc_airspeed_int_error(int16_t aspdError, int32_t aspd_integral);
 
 int 	airspeed		= 0;
 int 	groundspeed		= 0;
@@ -56,7 +56,7 @@ int cruise_airspeed			= CRUISE_AIRSPEED * 100;
 /// Filtered airspeed for less lumpiness
 int aspd_3DIMU_filtered = 0;	
 
-inline int get_filtered_airspeed( void ) {return  aspd_3DIMU_filtered;};
+inline int16_t get_filtered_airspeed( void ) {return  aspd_3DIMU_filtered;};
 
 void airspeedCntrl(void)
 {
@@ -77,10 +77,10 @@ void airspeedCntrl(void)
 
 // Calculate the required airspeed in cm/s.  desiredSpeed is in dm/s
 // airspeed and groundspeed in cm/s
-int calc_target_airspeed(int desiredSpd, unsigned int airspeed, unsigned int groundspeed)
+int calc_target_airspeed(int16_t desiredSpd, uint16_t airspeed, uint16_t groundspeed)
 {
 	union longww accum ;
-	int target;
+	int16_t target;
 
 	accum.WW = __builtin_mulsu ( desiredSpd , 10 ) ;
 	target = accum._.W0 ;
